fix remaining -Wextra enum/non-enum ternary warnings
Modern GCC's -Wextra flags 'enumerated and non-enumerated type in
conditional expression' for ternaries whose branches mix an enum
with an int/uint. Fix at source with explicit casts:

  Darkel.cpp      PedsKilledOfThisType index
  Restart.cpp     Hospital/Police override level vs curlevel (x2)
  Frontend.cpp    prevPage fallback page
  Font.h          FONT_LOCALE macro branches
  AudioLogic.cpp  Max() with VEHICLE_ONE_SHOT_CAR_JUMP_VOLUME
  CarAI.cpp       Max() with DRIVINGSTYLE_AVOID_CARS

Also bump the librw submodule for its matching warning fixes
(gl3device alpha ternaries, ps2raster %llX) and the new
LIBRW_STRICT_WARNINGS option. Native Linux GL3 and armhf GBM+SPI
cross builds now compile with zero warnings under -Wall -Wextra
-Wpedantic (RE3_STRICT_WARNINGS=OFF).

+++ src/audio/AudioLogic.cpp
@@ -2442,7 +2442,7 @@ cAudioManager::ProcessVehicleOneShots(cVehicleParams& params)
 		case SOUND_CAR_JUMP:
 		{
 			static uint8 iWheelIndex = 82;
-			Vol = Max(VEHICLE_ONE_SHOT_CAR_JUMP_VOLUME, 2 * (100 * m_asAudioEntities[m_sQueueSample.m_nEntityIndex].m_afVolume[i]));
+			Vol = Max((int)VEHICLE_ONE_SHOT_CAR_JUMP_VOLUME, 2 * (100 * m_asAudioEntities[m_sQueueSample.m_nEntityIndex].m_afVolume[i]));
 			maxDist = SQR(VEHICLE_ONE_SHOT_CAR_JUMP_MAX_DIST);
 			m_sQueueSample.m_nSampleIndex = SFX_TYRE_BUMP;
 			m_sQueueSample.m_nBankIndex = SFX_BANK_0;
